+++ src/griffe_fastapi/__init__.py
@@ -3,6 +3,15 @@
 Griffe extension for FastAPI.
 """
 
+from pathlib import Path
+
 from griffe_fastapi._extension import FastAPIExtension
 
-__all__: list[str] = ["FastAPIExtension"]
+
+def get_templates_path() -> Path:
+    """Return the templates directory path."""
+    print("Hey!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
+    return Path(__file__).parent / "templates"
+
+
+__all__: list[str] = ["get_templates_path", "FastAPIExtension"]