+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.4](https://github.com/fbraem/griffe-fastapi/releases/tag/0.1.2) - 2024-11-22
+
+- Feat: add api to the extra dictionary
+- Feat: add fastapi template
+
 ## [0.1.3](https://github.com/fbraem/griffe-fastapi/releases/tag/0.1.2) - 2024-11-19
 
 - Fix generated table in docstring.